// ui/src/generated/ProjectPatchResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProjectSnapshot } from "./ProjectSnapshot";

export type ProjectPatchResult = { before: ProjectSnapshot, after: ProjectSnapshot, affected_paths: Array<string>, changed_paths: Array<string>, git_exit_code: number | null, diagnostic: string, committed_to_git: boolean, };